+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["создан"]
+  "cSpell.words": ["изменен", "создан"]
 }
