+++ src/app/app.config.ts
@@ -4,18 +4,18 @@ import { routes } from './app.routes';
 import { provideHttpClient } from '@angular/common/http';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 import { provideStore } from '@ngrx/store';
-import { userReducer } from './user-list/store/users.reducer';
-import { UserEffects } from './user-list/store/user.effects';
 import { provideStoreDevtools } from '@ngrx/store-devtools';
-import { todoReducer } from './todo-list/store/todo.reducer';
 import { provideEffects } from '@ngrx/effects';
-import { TodoEffects } from './todo-list/store/todo.effects';
+import {userReducer} from "./components/user-list/store/users.reducer";
+import {todoReducer} from "./components/todo-list/store/todo.reducer";
+import {UserEffects} from "./components/user-list/store/user.effects";
+import {TodoEffects} from "./components/todo-list/store/todo.effects";
 
 export const appConfig: ApplicationConfig = {
   providers: [
     provideRouter(routes),
     provideHttpClient(),
-    provideAnimationsAsync(), 
+    provideAnimationsAsync(),
     provideStore({
       users: userReducer,
       todos: todoReducer
