+++ src/app/app.routes.ts
@@ -1,19 +1,19 @@
 import { Routes } from '@angular/router';
-import { MainComponent } from './main/main.component';
-import { TodosListComponent } from './todo-list/todos-list.component'; 
-import { authGuard } from './auth/auth.guard';
+import { MainComponent } from './components/main/main.component';
+import { authGuard } from './guards/auth/auth.guard';
+import {TodosListComponent} from "./components/todo-list/todos-list.component";
 
 
 export const routes: Routes = [
     { path: '', component: MainComponent },
     {
         path: 'admin',
-        loadComponent: () => import('./admin/admin.component').then(m => m.AdminComponent),
+        loadComponent: () => import('./components/admin/admin.component').then(m => m.AdminComponent),
         canActivate: [authGuard]
     },
     {
         path: 'users',
-        loadComponent: () => import('./user-list/user-list.component').then(m => m.UsersListComponent),
+        loadComponent: () => import('./components/user-list/user-list.component').then(m => m.UsersListComponent),
         canActivate: [authGuard]
     },
     { path: 'todos', component: TodosListComponent },
