+++ src/app/components/admin/login-dialog/login-dialog.component.scss
