+++ src/app/components/footer/footer.component.html
@@ -2,10 +2,10 @@
     <div class="footer container">
         <div class="footer__upper-block">
             <div class="footer__left-block">
-                <img class="company-logo" src="./assets/img/company-logo.svg" alt="company-logo">
+                <img class="company-logo" src="../../../assets/img/company-logo.svg" alt="company-logo">
                 <div class="contacts">
                     <div class="contacts__item">
-                        <img class="contacts__item-logo" src="./assets/img/call-calling.svg" alt="call calling">
+                        <img class="contacts__item-logo" src="../../../assets/img/call-calling.svg" alt="call calling">
                         <div class="contacts__item--gap">
                             <div class="contacts__item-text">Контактный телефон</div>
                             <div class="contacts__item-number">+ 7 (812) 309-09-34</div>
@@ -13,7 +13,7 @@
                         </div>
                     </div>
                     <div class="contacts__item">
-                        <img class="contacts__item-logo" src="./assets/img/sms.svg" alt="sms">
+                        <img class="contacts__item-logo" src="../../../assets/img/sms.svg" alt="sms">
                         <div class="contacts__item--gap">
                             <div class="contacts__item-text">Электронная почта</div>
                             <div class="contacts__item-number">tustin78&#64;mail.ru</div>
@@ -32,7 +32,7 @@
                     </div>
                     <div class="feedback-search">
                         <button class="feedback-button">
-                            <img class="contacts__item-logo" src="./assets/img/sms.svg" alt="sms">
+                            <img class="contacts__item-logo" src="../../../assets/img/sms.svg" alt="sms">
                             <input class="feedback-search-item" type="search" placeholder="Введите свою почту">
                             <button class="feedback-button-send">
                                 <span class="feedback-button-send-text">Отправить</span>
@@ -50,9 +50,9 @@
             </nav>
             <div class="footer__lower-block-nav">
                 <div class="social-media">
-                    <a href=""><img src="./assets/img/facebook.svg" alt="facebook"></a>
-                    <a href=""><img src="./assets/img/whatsapp.svg" alt="whatsapp"></a>
-                    <a href=""><img src="./assets/img/youtube.svg" alt="youtube"></a>
+                    <a href=""><img src="../../../assets/img/facebook.svg" alt="facebook"></a>
+                    <a href=""><img src="../../../assets/img/whatsapp.svg" alt="whatsapp"></a>
+                    <a href=""><img src="../../../assets/img/youtube.svg" alt="youtube"></a>
                 </div>
                 <div class="footer--text" style="color: #FFFFFF99;">© 2021-2022, EDE</div>
             </div>
