+++ src/app/components/footer/footer.component.scss
@@ -2,8 +2,8 @@
     height: 463px;
     background-color: #4B565E;
     margin-top: 48px;
-    background-image: url('/assets/img/line-horizontal.png'),
-        url('/assets/img/line-vertical.png');
+    background-image: url('/src/assets/img/line-horizontal.png'),
+        url('/src/assets/img/line-vertical.png');
     background-repeat: no-repeat;
     background-position: center center, top center;
 }
