+++ src/app/components/header/header.component.scss
@@ -23,7 +23,6 @@
 }
 
 .header-navigation-button {
-    padding: 16px;
     display: flex;
     gap: 16px;
     justify-content: center;
@@ -209,7 +208,7 @@
 }
 
 .button-styles-forms {
-    padding: 7px;
+    padding: 6px;
     font-size: 16px;
     border: 1px solid #F0BA4E;
     background: none;
