+++ src/app/components/header/header.component.ts
@@ -1,13 +1,13 @@
-import { DatePipe, NgFor, NgIf } from '@angular/common';
-import { Component, inject } from '@angular/core';
-import { RouterLink } from '@angular/router';
-import { YellowDirective } from '../../directives/yellow.directive';
-import { MatButtonModule } from '@angular/material/button';
-import { MatDialog } from '@angular/material/dialog';
-import { LoginDialogComponent } from '../../admin/login-dialog/login-dialog.component';
-import { Store } from '@ngrx/store';
-import { CommonModule } from '@angular/common';
-import { selectCurrentUser, selectIsAdmin, selectIsLoggedIn } from '../../user-list/store/users.selectors';
+import {DatePipe, NgFor, NgIf} from '@angular/common';
+import {Component, inject} from '@angular/core';
+import {RouterLink} from '@angular/router';
+import {YellowDirective} from '../../directives/yellow.directive';
+import {MatButtonModule} from '@angular/material/button';
+import {MatDialog} from '@angular/material/dialog';
+import {Store} from '@ngrx/store';
+import {CommonModule} from '@angular/common';
+import {LoginDialogComponent} from "../admin/login-dialog/login-dialog.component";
+import {selectCurrentUser, selectIsAdmin, selectIsLoggedIn, selectUsers} from "../user-list/store/users.selectors";
 
 const aboutCompanyFn = (text: string) => text;
 
@@ -49,9 +49,6 @@ export class HeaderComponent {
     this.menuItems = this.menuItems.map(
       (item: string) => this.isUpperCase ? item.toLowerCase() : item.toUpperCase()
     )
-
     this.isUpperCase = !this.isUpperCase;
-
-
   }
 }
