+++ src/app/components/user-list/create-edit-user/create-edit-user.component.html
@@ -1,0 +1,32 @@
+<!--<div class="edit">-->
+<!--  <h1 class="edit-title">EDIT USER FORM</h1>-->
+<!--  <div class="edit-user">-->
+<!--    <form class="form" [formGroup]="form">-->
+<!--      <mat-form-field class="edit-user-form">-->
+<!--        <mat-label>Введите имя пользователя</mat-label>-->
+<!--        <input matInput formControlName="name" placeholder="name">-->
+<!--        <mat-error>Некорректное название</mat-error>-->
+<!--      </mat-form-field>-->
+<!--      <mat-form-field class="edit-user-form">-->
+<!--        <mat-label>Введите email</mat-label>-->
+<!--        <input matInput formControlName="email" placeholder="email">-->
+<!--        <mat-error>Некорректное название</mat-error>-->
+<!--      </mat-form-field>-->
+<!--      <mat-form-field class="edit-user-form">-->
+<!--        <mat-label>Введите website</mat-label>-->
+<!--        <input matInput formControlName="website" placeholder="website">-->
+<!--        <mat-error>Некорректное название</mat-error>-->
+<!--      </mat-form-field>-->
+<!--      <mat-form-field class="edit-user-form">-->
+<!--        <mat-label>Введите company</mat-label>-->
+<!--        <input matInput formControlName="companyName" placeholder="company name">-->
+<!--        <mat-error>Некорректное название</mat-error>-->
+<!--      </mat-form-field>-->
+<!--      <button matTooltip="Редактировать данные пользователя" class="edit-user-form" matIconSuffix-->
+<!--              mat-raised-button type="submit" [disabled]="form.invalid" [mat-dialog-close]="userWithUpdateFields">-->
+<!--        Edit user-->
+<!--        <mat-icon>add</mat-icon>-->
+<!--      </button>-->
+<!--    </form>-->
+<!--  </div>-->
+<!--</div>-->
