+++ src/app/components/user-list/create-edit-user/create-edit-user.component.ts
@@ -1,0 +1,27 @@
+import { Component } from '@angular/core';
+import {MatButton} from "@angular/material/button";
+import {MatError, MatFormField, MatLabel, MatSuffix} from "@angular/material/form-field";
+import {MatIcon} from "@angular/material/icon";
+import {MatInput} from "@angular/material/input";
+import {MatTooltip} from "@angular/material/tooltip";
+import {ReactiveFormsModule} from "@angular/forms";
+
+@Component({
+  selector: 'app-create-edit-user',
+  imports: [
+    MatButton,
+    MatError,
+    MatFormField,
+    MatIcon,
+    MatInput,
+    MatLabel,
+    MatSuffix,
+    MatTooltip,
+    ReactiveFormsModule
+  ],
+  templateUrl: './create-edit-user.component.html',
+  styleUrl: './create-edit-user.component.scss'
+})
+export class CreateEditUserComponent {
+
+}
