+++ src/app/components/user-list/create-user-dialog/create-user-form/create-user-form.component.ts
@@ -5,7 +5,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MAT_DIALOG_DATA, MatDialogClose } from "@angular/material/dialog";
-import { User } from '../../../interface/users';
+import {User} from "../../../../interface/users";
 
 @Component({
     selector: 'app-create-user-form',
