+++ src/app/components/user-list/edit-user-dialog/edit-user-dialog.component.ts
@@ -6,7 +6,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MAT_DIALOG_DATA, MatDialogClose } from '@angular/material/dialog';
-import { User } from '../../interface/users';
+import {User} from "../../../interface/users";
 
 @Component({
     selector: 'app-edit-user-dialog',
