+++ src/app/components/user-list/user-card/delete-user-dialog/delete-user-dialog.component.ts
@@ -5,7 +5,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MAT_DIALOG_DATA, MatDialogActions, MatDialogContent, MatDialogRef } from '@angular/material/dialog';
-import { User } from '../../../interface/users';
+import {User} from "../../../../interface/users";
 
 @Component({
     selector: 'app-delete-user-dialog',
