+++ src/app/components/user-list/user-card/user-card.component.scss
